refactor(mcp-server): split proxy pdp e2e tests
